Avoid g_type_init() deprecation errors for Vala souurces
Inject g_type_init() redefinition into C source generated from Vala.
This is needed as the current GLib dependency is 2.32, which includes
g_type_init(), is also used as the Vala target GLib version. Therefore,
the C sources generated from the Vala sources include g_type_init()
calls. As it was deprecated in GLib 2.35.0, this creates a warning
(which becomes an error and stops the build) when compiling against that
version or higher.